requirements.txt
accelerate==1.11.0
annotated-types==0.7.0
anyio==4.11.0
black==25.11.0
blockbuster==1.5.25
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==44.0.3
cycler==0.12.1
distro==1.9.0
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
forbiddenfruit==0.1.4
fsspec==2025.10.0
google-auth==2.47.0
google-genai==1.57.0
googleapis-common-protos==1.72.0
grandalf==0.8
grpcio==1.76.0
grpcio-tools==1.75.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
kiwisolver==1.4.9
langchain-core==1.2.7
langchain-google-genai==4.1.3
langgraph==1.0.5
langgraph-api==0.5.13
langgraph-checkpoint==3.0.1
langgraph-cli==0.4.7
langgraph-prebuilt==1.0.2
langgraph-runtime-inmem==0.18.0
langgraph-sdk==0.3.2
langsmith==0.4.42
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orchestrator==1.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.1
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.50.0
structlog==25.5.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
truststore==0.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
utils==1.0.2
uuid_utils==0.13.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
